Next runs

+++ print_r2_values.py
@@ -1,6 +1,7 @@
 import pickle
 import pandas as pd
 import os
+import numpy as np
 
 file_path = '_current_run_saved_files'
 dict_run = {}
@@ -12,12 +13,12 @@ for folder in os.listdir(file_path):
             dict_run[folder[-2:]] = {'run_no': int(folder[-2:]), 'r2_train': dict_hp['r2 train'],
                                      'r2_valid': dict_hp['r2 valid'],
                                      'difference': dict_hp['r2 train'] - dict_hp['r2 valid'],
-                                     'lambda': dict_hp['regularization factor']}
+                                     'lambda': np.float(dict_hp['regularization factor'])}
         except:
             dict_run[folder[-1]] = {'run_no': int(folder[-1]), 'r2_train': dict_hp['r2 train'],
                                      'r2_valid': dict_hp['r2 valid'],
                                      'difference': dict_hp['r2 train'] - dict_hp['r2 valid'],
-                                     'lambda': dict_hp['regularization factor']}
+                                     'lambda': np.float(dict_hp['regularization factor'])}
     except:
         print('Folder name: ', folder, ' is not a run')
 print('=====================================================================')
